backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
websockets>=10.0
pydub>=0.25.1
librosa>=0.10.0
numpy>=1.21.0
openai-whisper>=20230918
nltk>=3.6.3
spacy>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0 